// src/modules/mailing/api/MailingApiRoutes.ts
export enum MailingApiRoutes {
  // mailing
  GET_MAILBOX_INFO = '/api/mailing/mailboxes',
  GET_SECTION_MAIL_THREAD_INFO = '/api/mailing/section/:type/messages',
  GET_MAILBOX_MAIL_THREAD_INFO = '/api/mailing/mailboxes/:mailboxId/messages',
  GET_MAIL_MESSAGES = '/api/mailing/mailboxes/:mailboxId/threads/:messageId',
  DOWNLOAD_FILE_FROM_MESSAGE = '/api/mailing/mailboxes/:mailboxId/messages/:messageId/attachments/:payloadId',
  SEND_MESSAGE = '/api/mailing/mailboxes/:mailboxId/send',
  SPAM_THREAD = '/api/mailing/mailboxes/:mailboxId/threads/:messageId/spam',
  UNSPAM_THREAD = '/api/mailing/mailboxes/:mailboxId/threads/:messageId/unspam',
  TRASH_THREAD = '/api/mailing/mailboxes/:mailboxId/threads/:messageId/trash',
  UNTRASH_THREAD = '/api/mailing/mailboxes/:mailboxId/threads/:messageId/untrash',
  SEEN_THREAD = '/api/mailing/mailboxes/:mailboxId/threads/:messageId/seen',
  UNSEEN_THREAD = '/api/mailing/mailboxes/:mailboxId/threads/:messageId/unseen',
  MAIL_CREATE_CONTACT = '/api/mailing/mailboxes/:mailboxId/messages/:messageId/contact',
  GET_MAIL_MESSAGE = '/api/mailing/mailboxes/:mailboxId/messages/:messageId',
  // mailbox signature
  GET_SIGNATURES = '/api/mailing/signatures',
  ADD_SIGNATURE = '/api/mailing/signatures',
  UPDATE_SIGNATURE = '/api/mailing/signatures/:id',
  DELETE_SIGNATURE = '/api/mailing/signatures/:id',
  // mailing settings
  SEND_FEEDBACK = '/api/mailing/feedback',
  GET_MAILBOXES_SETTINGS = '/api/mailing/settings/mailboxes',
  ADD_MAILBOX_SETTINGS = '/api/mailing/settings/mailboxes',
  UPDATE_MAILBOX_SETTINGS = '/api/mailing/settings/mailboxes/:id',
  DELETE_MAILBOX_SETTINGS = '/api/mailing/settings/mailboxes/:id',
  GMAIL_CONNECT_SETTINGS = '/api/mailing/settings/mailboxes/gmail/connect/:id',
  GET_MAILBOX_MANUAL_SETTINGS = '/api/mailing/settings/mailboxes/:id/manual',
  UPDATE_MAILBOX_MANUAL_SETTINGS = '/api/mailing/settings/mailboxes/:id/manual',
}